Add escaping when inserting CSS
This also moves the handler insertion back to the loader.
