+++ packages/runtime-common/package.json
@@ -32,6 +32,7 @@
     "flat": "^5.0.2",
     "http-status-codes": "^2.2.0",
     "ignore": "^5.2.0",
+    "js-string-escape": "^1.0.1",
     "json-typescript": "^1.1.2",
     "lodash": "^4.17.21",
     "loglevel": "^1.8.1",
@@ -44,6 +45,7 @@
   "devDependencies": {
     "@babel/core": "^7.17.8",
     "@types/dompurify": "^3.0.2",
+    "@types/js-string-escape": "^1.0.1",
     "@types/qunit": "^2.11.3",
     "dompurify": "^3.0.3",
     "ember-cli-htmlbars": "^6.2.0"
